[lsp] LSP server implementation
This is a contribution from Arash Sahebolamri,
adapted to fit in a new project. Original PR is

	https://github.com/AlloyTools/org.alloytools.alloy/pull/162

It was adapted in two PRs. This PR contains the
pure server code in a new project. The other
contains changes to the core that were needed by
this project. 

The changes that were added were:

- Alloy main class now dispatches to either
  the SimplGUI or the AlloyLanguageServer
- Some of the submitted classes contained more than
  one class in one file. Nesting is fine, but this
  can be awkward. They were separated.
- Formatted to fit the standard